Improved the navigation
Changed the main window so that it has a current view which is initially
set to "all lists" (this used to be all in the main window).

Each "present list" is now a button rather than a textblock so you can
handle the button pressed in a more standard manner.  Clicking this
currently displays the same view but should take you to a list of all of
the items on the list.