Added Fluent NHibernate
Added Fluent NHibernate so created mapping classes and made changes to
the domain models to support Fluent NH.  Also plugged Fluent NHibernate
into Castle Windsor and hooked it up to a SQL Server Express install.

Tried to use the Castle.Facilities.AutoTx package to manage transactions
in the repositories but it is currently broken and doesn't produce a
DLL.  Also can't get older versions to work with Windsor.