Added eager loading of presents to lists
Presents are now eagerly loaded when the lists are loaded from the
database. This is used to show a count of the number of presents on the
list view screen.

Also removed the mock list repository and added a class for proper
implementation although this is currently empty as the presents are
eagerly loaded.

Also changed the map classes so that they include the actual name of the
database column. This was causing invalid SQL when joining the present
and lists tables.