Presents are now displayed on the list overview
Created a new view model and view to show the present description. One
of these is created for each present on a list and is displayed on the
presents overview view.

There is a bug which causes the lists to be displayed as many times as
there are presents. This is probably because of a cartisian in the SQL
to get the lists with the presents eagerly fetched.