Changed how the navigation works
Navigating between view models now uses specific classes for the action
rather than just ViewModelBase. This is to allow navigating backwards as
well as forwards (before you couldn't differentiate between the two).
When you press save on the EditableListViewModel you go back to the
previous page.

The implementation isn't ideal because you have to construct a
GoToPreviousViewModel whenever you want to go back and the
GoBackOneViewModel() method has a parameter that is unused.