Changed the appearance of the PresentListOverview
The PresentListOverviewView is now a panel with text on it rather than a
button.

The AllListsView has changed to use a listbox of
PresentListOverviewViews rather than a wrap panel inside an items
control.

Also changed the MainWindowView so that the row where the current view
is displayed has a row height of * (fill) rather than auto as this was
preventing the scrollbar appearing on the AllListsView.