Changed the PresentListView
Added the list notes to the view.
Added a border underneath the list name and to the left of the arrow.
Added padding between the lists on the all lists view.
Resizing the all lists view in the horizontal direction will make the
list name and notes truncate.
The PresentListOverview is no longer responsible for displaying the
presents, that is done when selecting a list on the all lists view.