Changed how the child window is displayed
Consolidated two views and view models (failed to save data and failed
to retrieve data) into one error view and view model. There are still
separate messages which the child window listens for. When it receives a
message it creates the error view model and determines the text for it
based on the message received. This probably needs to be moved to the
error message view model later